parametric study fixes

+++ scenarios/parametric_study/main.py
@@ -1,12 +1,9 @@
 import argparse
 import copy
 import json
-import logging
-import math
 import os
 import random
 import shutil
-import sys
 
 import numpy as np
 import pandas as pd
@@ -29,6 +26,7 @@ def main(lower_bound : int, upper_bound : int, level : int):
     # read scenario parameters file
     scenarios_file = os.path.join(scenario_dir, 'resources', 'lhs_scenarios.csv')
     scenarios_df : pd.DataFrame = pd.read_csv(scenarios_file)
+    scenarios_df = scenarios_df.sort_values(by=['num_planes','num_sats_per_plane'])
 
     # check if bounds are valid
     assert 0 <= lower_bound <= upper_bound
@@ -38,14 +36,14 @@ def main(lower_bound : int, upper_bound : int, level : int):
     # set fixed parameters
     same_events = True
     overwrite = False
-    pregenerated_events = False
+    sim_duration = 24.0 / 24.0 # in days
     preplanners = [
                         'naive',
                         'dynamic'
                         # 'nadir'                    
                         ]
     preplanner_settings = [
-        (100, 100),
+        (100, 100),         # period[s], horizon [s]
         (500, 500),
         (np.Inf, np.Inf)
     ]
@@ -54,7 +52,7 @@ def main(lower_bound : int, upper_bound : int, level : int):
                     'acbba',
                     ]
     bundle_sizes = [
-                    1,
+                    # 1,
                     # 2, 
                     3 
                     # 5
@@ -66,10 +64,10 @@ def main(lower_bound : int, upper_bound : int, level : int):
 
     # run simulation for each set of parameters
     with tqdm.tqdm(total=n_runs) as pbar:
-        for i,row in scenarios_df.iterrows():
-            if i < lower_bound:
+        for scenario_i,row in scenarios_df.iterrows():
+            if scenario_i < lower_bound:
                 continue
-            elif upper_bound < i:
+            elif upper_bound < scenario_i:
                 break
 
             # extract parameters
@@ -100,30 +98,48 @@ def main(lower_bound : int, upper_bound : int, level : int):
             max_severity = 100
             measurement_list = ['sar', 'visual', 'thermal']
 
-            sim_duration = 24.0 / 24.0
-
-            pregenerated_events = False
-
             # remove previous orbit data
-            clear_orbitdata(scenario_dir)
+            # clear_orbitdata(scenario_dir)
 
             # run cases
+            pregenerated_events = False
             for preplanner in preplanners:
                 for period, horizon in preplanner_settings:
                     for replanner in replanners:
                         for bundle_size in bundle_sizes:
                             # check simulation requirements
-                            if replanner == 'broadcaster' and bundle_size > 1:
+                            if replanner == 'broadcaster' and bundle_size > 1 and len(bundle_sizes) > 1:
                                 break # skip
 
-                            if preplanner == 'np' and horizon == np.Inf:
+                            if preplanner == 'dynamic' and horizon == np.Inf:
                                 continue # skip
 
                             if preplanner == 'naive' and horizon != np.Inf:
                                 continue # skip
 
+                            if preplanner == 'naive' and replanner == 'broadcaster':
+                                continue # skip
+
                             # create specs from template
                             scenario_specs : dict = copy.deepcopy(template_specs)
+
+                            # create scenario name
+                            scenario_name = f'{parent_scenario_name}_{scenario_i}_{preplanner}-{period}-{horizon}_{replanner}'
+                            if replanner != 'broadcaster': scenario_name += f'-{bundle_size}'
+                            
+                            # set scenario name
+                            scenario_specs['scenario']['name'] = scenario_name
+
+                            # set outdir
+                            orbitdata_dir = os.path.join('./scenarios', parent_scenario_name, 'orbit_data', f'{parent_scenario_name}_{scenario_i}')
+                            scenario_specs['settings']['outDir'] = orbitdata_dir
+
+                            # check overwrite toggle
+                            results_dir = os.path.join('./scenarios', parent_scenario_name, 'results', scenario_name)
+                            if not overwrite and os.path.exists(os.path.join(results_dir, 'summary.csv')):
+                                # scenario already ran; skip to avoid overwrite
+                                pbar.update(1)
+                                continue
 
                             # set simulation duration
                             scenario_specs['duration'] = sim_duration
@@ -131,22 +147,21 @@ def main(lower_bound : int, upper_bound : int, level : int):
                             # set events
                             if not (same_events and pregenerated_events):
                                 if event_clustering == 'uniform':
-                                    grid_path = create_uniform_grid(scenario_dir, n_events)
+                                    grid_path = create_uniform_grid(scenario_dir, scenario_i, n_events)
                                 elif event_clustering == 'clustered':
-                                    grid_path = create_clustered_grid(scenario_dir, n_events)
+                                    grid_path = create_clustered_grid(scenario_dir, scenario_i, n_events)
                                 else:
                                     raise NotImplementedError(f'event clustering of type {event_clustering} not supported.')
-
-                                create_events(scenario_dir, grid_path, sim_duration, n_events, event_duration, min_severity, max_severity, measurement_list)
+                                create_events(scenario_dir, scenario_i, grid_path, sim_duration, n_events, event_duration, min_severity, max_severity, measurement_list)
                                 pregenerated_events = True
                                 
                             scenario_specs['grid'] = [{
                                                         "@type": "customGrid",
-                                                        "covGridFilePath": os.path.join(scenario_dir, 'resources', f'random_grid.csv')
+                                                        "covGridFilePath": os.path.join(scenario_dir, 'resources', f'random_grid_{scenario_i}.csv')
                                                     }]
                             scenario_specs['scenario']['events'] = {
                                                         "@type": "PREDEF", 
-                                                        "eventsPath" : os.path.join(scenario_dir, 'resources', f'random_events.csv')
+                                                        "eventsPath" : os.path.join(scenario_dir, 'resources', f'random_events_{scenario_i}.csv')
                                                     }
 
                             # set spacecraft specs
@@ -188,6 +203,9 @@ def main(lower_bound : int, upper_bound : int, level : int):
                                     # set replanner
                                     sat['planner']['replanner']['@type'] = replanner
                                     sat['planner']['replanner']['bundle size'] = bundle_size
+
+                                    # set science
+                                    sat['science']['eventsPath'] = os.path.join(scenario_dir, 'resources', f'random_events_{scenario_i}.csv')
                                     
                                     # add to list of sats
                                     sats.append(sat)
@@ -198,27 +216,13 @@ def main(lower_bound : int, upper_bound : int, level : int):
                             # update list of satellites
                             scenario_specs['spacecraft'] = sats
 
-                            # create scenario name
-                            scenario_name = f'{parent_scenario_name}_{i}_{preplanner}-{period}-{horizon}_{replanner}'
-                            if replanner != 'broadcaster': scenario_name += f'-{bundle_size}'
-                            
-                            # set scenario name
-                            scenario_specs['scenario']['name'] = scenario_name
-
-                            # check overwrite toggle
-                            results_dir = os.path.join('./scenarios',parent_scenario_name, 'results', scenario_name)
-                            if not overwrite and os.path.exists(os.path.join(results_dir, 'summary.csv')):
-                                # scenario already ran; skip to avoid overwrite
-                                pbar.update(1)
-                                continue
+                            # print welcome message
+                            print_welcome(scenario_name)
 
                             # initialize mission
                             mission : Mission = Mission.from_dict(scenario_specs)
 
-                            # print welcome message
-                            print_welcome(scenario_name)
-
-                            # execute mission
+                            # # execute mission
                             mission.execute()
                                 
                             # update progress bad
@@ -238,41 +242,50 @@ def clear_orbitdata(scenario_dir : str) -> None:
                     shutil.rmtree(file_path)
             except Exception as e:
                 print('Failed to delete %s. Reason: %s' % (file_path, e))
-        shutil.rmtree(orbitdata_path)
 
-def create_uniform_grid(scenario_dir, n_events : int, lat_spacing : float = 0.1, lon_spacing : float = 0.1) -> str:
+def create_uniform_grid(scenario_dir : str, scenario_i : str, n_events : int, lat_spacing : float = 0.1, lon_spacing : float = 0.1) -> str:
+    # set grid name
+    grid_path : str = os.path.join(scenario_dir, 'resources', f'random_grid_{scenario_i}.csv')
+    
+    # check if grid already exists
+    if os.path.isfile(grid_path): return grid_path
+
+    # generate grid
     all_groundpoints = [(lat, lon) 
                         for lat in np.linspace(-90, 90, int(180/lat_spacing)+1)
                         for lon in np.linspace(-180, 180, int(360/lon_spacing)+1)
                         if lon < 180
                         ]
-    
+
     groundpoints : list = random.sample(all_groundpoints, n_events)
-    # groundpoints = [(np.round(lat,1),np.round(lon,1))
-    #                 for lat,lon in groundpoints]
     groundpoints.sort()
     
     # create dataframe
     df = pd.DataFrame(data=groundpoints, columns=['lat [deg]','lon [deg]'])
 
     # save to csv
-    grid_path : str = os.path.join(scenario_dir, 'resources', f'random_grid.csv')
     df.to_csv(grid_path,index=False)
 
     # return address
     return grid_path
 
-def create_clustered_grid(scenario_dir, n_events : int, variance : float = 1.0, n_clusters : float = 100) -> str:
+def create_clustered_grid(scenario_dir : str, scenario_i : str, n_events : int, variance : float = 1.0, n_clusters : float = 100) -> str:
+    # set grid name
+    grid_path : str = os.path.join(scenario_dir, 'resources', f'random_grid_{scenario_i}.csv')
+    
+    # check if grid already exists
+    if os.path.isfile(grid_path): return grid_path
+    
     # create clustered grid of gound points
     std = np.sqrt(variance)
     groundpoints = []
-    
-    for _ in range(n_clusters):
+
+    for _ in tqdm.tqdm(range(n_clusters), desc='generating clustered grid', leave=False):
         # find cluster center
         lat_cluster = (90 - -90) * random.random() -90
         lon_cluster = (180 - -180) * random.random() -180
         
-        for _ in range(n_events):
+        for _ in range(int(n_events / n_clusters)):
             # sample groundpoint
             lat = random.normalvariate(lat_cluster, std)
             lon = random.normalvariate(lon_cluster, std)
@@ -282,19 +295,24 @@ def create_clustered_grid(scenario_dir, n_events : int, variance : float = 1.0, 
     df = pd.DataFrame(data=groundpoints, columns=['lat [deg]','lon [deg]'])
 
     # save to csv
-    grid_path : str = os.path.join(scenario_dir, 'resources', f'random_grid.csv')
     df.to_csv(grid_path,index=False)
 
     # return address
     return grid_path
 
-def create_events(scenario_dir : str, grid_path : str, sim_duration : float, n_events : int, event_duration : float, min_severity : float, max_severity : float, measurements : list) -> str:
+def create_events(scenario_dir : str, scenario_i : str, grid_path : str, sim_duration : float, n_events : int, event_duration : float, min_severity : float, max_severity : float, measurements : list) -> str:
+    # set events path
+    events_path = os.path.join(scenario_dir, 'resources', f'random_events_{scenario_i}.csv')
+    
+    # check if events have already been generated
+    if os.path.isfile(events_path): return events_path
+
     # load coverage grid
     grid : pd.DataFrame = pd.read_csv(grid_path)
 
     # generate events
     events = []
-    for _ in range(n_events):
+    for _ in tqdm.tqdm(range(n_events), desc='generating events', leave=False):
         # select ground points for events
         gp_index = random.randint(0, len(grid)-1)
         gp = grid.iloc[gp_index]
@@ -330,13 +348,10 @@ def create_events(scenario_dir : str, grid_path : str, sim_duration : float, n_e
         # add to list of events
         events.append(event)
 
-    assert all([(lat,lon) in grid.values for lat,lon,*_ in events])
-
     # compile list of events
     events_df = pd.DataFrame(data=events, columns=['lat [deg]','lon [deg]','start time [s]','duration [s]','severity','measurements'])
     
     # save list of events to events path 
-    events_path = os.path.join(scenario_dir, 'resources', 'random_events.csv')
     events_df.to_csv(events_path,index=False)
 
     # return path address
